Update manifest.json
Fixes issue #21. Add the now required version attribute to the component manifest.json. Tested change locally and confirmed the integration lists and loads again. Note that if you are going to cut a new release to include this change, you'll want to reflect the updated version in the manifest file!